+++ issues/689/.gitignore
@@ -1,2 +1,2 @@
 *.class
-node_modules
+target
