+++ issues/689/src/CopyAndPasteAnImage.java
