+++ issues/689/src/ThreadClassDemo.java
