+++ issues/689/src/jdbcConn.java
